backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
python-jose[cryptography]==3.3.0
supabase==2.3.4
openai==1.6.1
requests==2.31.0
yt-dlp==2023.12.30
python-dotenv==1.0.0
aiofiles==23.2.1
ffmpeg-python==0.2.0